refactor(ai-tools): extract getEffectiveLimit + executeCountOperation to tool-executor-helpers
