+++ nodes/Autotask/ai-tools/tool-executor.ts
@@ -49,7 +49,6 @@ import {
 	dispatchOperationResponse,
 	MAX_RESPONSE_RECORDS,
 } from './operation-handlers/operation-dispatch';
-import { CountOperation } from '../operations/base/count-operation';
 import type { IAutotaskEntity } from '../types';
 import {
 	buildFieldLookup,
@@ -287,16 +286,9 @@ export async function resolveCompanyToProjectIdFilter(
 	};
 }
 
-export const DEFAULT_QUERY_LIMIT = 10;
-export const MAX_QUERY_LIMIT = 500;
+import { DEFAULT_QUERY_LIMIT, MAX_QUERY_LIMIT, getEffectiveLimit, executeCountOperation } from './tool-executor-helpers';
+export { DEFAULT_QUERY_LIMIT, MAX_QUERY_LIMIT, getEffectiveLimit, executeCountOperation };
 export const RECENCY_OVER_REQUEST_LIMIT = 500;
-
-export function getEffectiveLimit(limit: number | undefined): number {
-	if (typeof limit !== 'number' || Number.isNaN(limit)) {
-		return DEFAULT_QUERY_LIMIT;
-	}
-	return Math.min(Math.max(Math.trunc(limit), 1), MAX_QUERY_LIMIT);
-}
 
 /**
  * Build field values for create/update from params.
@@ -497,33 +489,6 @@ function buildContractViolationNextAction(
 	);
 }
 
-// Used for count-injection. Must NOT route through executeToolOperation — the two would
-// share (and race on) the same context.getNodeParameter override.
-async function executeCountOperation(
-	resource: string,
-	filters: unknown[],
-	context: IExecuteFunctions,
-): Promise<number | null> {
-	try {
-		const scopedContext = Object.create(context) as IExecuteFunctions;
-		scopedContext.getNodeParameter = ((
-			name: string,
-			_index: number,
-			fallback?: unknown,
-		): unknown => {
-			if (name === 'filtersFromTool') return filters;
-			if (name === 'returnAll') return false;
-			if (name === 'id') return null;
-			return context.getNodeParameter(name, 0, fallback);
-		}) as IExecuteFunctions['getNodeParameter'];
-		const countOp = new CountOperation<IAutotaskEntity>(resource, scopedContext);
-		return await countOp.execute(0);
-	} catch (error) {
-		const message = error instanceof Error ? error.message : String(error);
-		console.debug('[executeCountOperation] count call failed:', message);
-		return null;
-	}
-}
 
 /**
  * Execute an Autotask operation by routing to the existing tool executor
